@@ -2,8 +2,22 @@
     "name": "svg-spritemap-webpack-plugin",
     "version": "4.7.0",
     "description": "Generates symbol-based SVG spritemap from all .svg files in a directory",
-    "main": "lib/index.js",
     "homepage": "https://github.com/cascornelissen/svg-spritemap-webpack-plugin",
+    "license": "MIT",
+    "main": "index.js",
+    "type": "module",
+    "types": "index.d.ts",
+    "exports": {
+        ".": {
+            "types": "./index.d.ts",
+            "default": "./index.js"
+        }
+    },
+    "author": {
+        "name": "Cas Cornelissen",
+        "email": "cas.cornelissen@onefinity.io",
+        "url": "https://www.onefinity.io/"
+    },
     "repository": {
         "type": "git",
         "url": "git+https://github.com/cascornelissen/svg-spritemap-webpack-plugin.git"
@@ -23,49 +37,65 @@
         "webpack",
         "plugin"
     ],
-    "author": {
-        "name": "Cas Cornelissen",
-        "email": "cas.cornelissen@onefinity.io",
-        "url": "https://www.onefinity.io/"
-    },
-    "license": "MIT",
-    "files": [
-        "lib/",
-        "docs/"
-    ],
     "scripts": {
-        "postversion": "npm run test:coverage && codecov",
-        "test": "jest --verbose",
-        "test:coverage": "jest --collectCoverageFrom='lib/**/*.js' --collectCoverageFrom='!lib/templates/**' --coverage"
+        "build": "tsc",
+        "lint": "eslint '**/*.ts' --ignore-pattern '**/*.d.ts'",
+        "test": "tsx --experimental-test-coverage --test 'source/test.ts'",
+        "prepack": "npm run lint && npm run test && npm run build",
+        "postpack": "rimraf --glob 'index.{js,d.ts}' 'source/*.{js,d.ts}' 'source/helpers/**/*.{js,d.ts}'",
+        "prepare": "husky"
     },
+    "files": [
+        "index.js",
+        "index.d.ts",
+        "source/**/*.js",
+        "source/**/*.d.ts"
+    ],
     "engines": {
-        "node": ">= 10.13.0"
+        "node": "^20.11.0 || ^21.2.0 || >= 22.0.0"
     },
     "dependencies": {
-        "@xmldom/xmldom": "^0.8.2",
-        "glob": "^8.0.3",
-        "joi": "^17.6.0",
-        "loader-utils": "^3.2.0",
-        "lodash": "^4.17.21",
+        "@xmldom/xmldom": "^0.9.8",
+        "glob": "^11.0.3",
+        "loader-utils": "^3.3.1",
+        "lodash-es": "^4.17.21",
         "mini-svg-data-uri": "^1.4.4",
-        "mkdirp": "^1.0.4",
-        "svg-element-attributes": "^1.3.1",
-        "svg4everybody": "^2.1.9",
-        "svgo": "^2.8.0",
-        "webpack-merge": "^5.8.0",
-        "webpack-sources": "^3.2.3"
+        "mkdirp": "^3.0.1",
+        "svg-element-attributes": "^2.1.0",
+        "webpack-merge": "^6.0.1",
+        "webpack-sources": "^3.3.3",
+        "zod": "^4.0.10"
     },
     "devDependencies": {
-        "@babel/preset-env": "^7.18.2",
-        "babel-jest": "^28.1.1",
-        "codecov": "^3.8.3",
-        "cross-env": "^7.0.3",
-        "del": "^6.1.1",
-        "jest": "^28.1.1",
-        "webpack": "^5.73.0",
-        "webpack-cli": "^4.10.0"
+        "@babel/preset-env": "^7.28.0",
+        "@onefinity/eslint-config": "^3.6.1",
+        "@types/loader-utils": "^2.0.6",
+        "@types/lodash-es": "^4.17.12",
+        "cross-env": "^10.0.0",
+        "css-loader": "^7.1.2",
+        "del": "^8.0.0",
+        "html-webpack-plugin": "^5.6.3",
+        "husky": "^9.1.7",
+        "less": "^4.4.0",
+        "less-loader": "^12.3.0",
+        "mini-css-extract-plugin": "^2.9.2",
+        "rimraf": "^6.0.1",
+        "sass": "^1.89.2",
+        "sass-loader": "^16.0.5",
+        "tsx": "^4.20.3",
+        "type-fest": "^4.41.0",
+        "webpack": "^5.101.0",
+        "webpack-cli": "^6.0.1"
     },
     "peerDependencies": {
+        "svg4everybody": "^2.1.9",
+        "svgo": "^4.0.0",
+        "typescript": "^4.0.0 || ^5.0.0",
         "webpack": "^5.0.0"
+    },
+    "peerDependenciesMeta": {
+        "typescript": {
+            "optional": true
+        }
     }
 }
