@@ -1,1 +1,1 @@
-require('./scss/style.scss');
+import './scss/style.scss';
